Create client.c
Code Description: SCTP Client

This C program represents an SCTP (Stream Control Transmission Protocol) client. SCTP is a communication protocol that combines some advantages of both TCP and UDP. This client establishes a connection to a remote server and allows the user to send messages to the server.

Key Functionalities of the Code:

The client creates a socket for communication.
It establishes a connection with a remote server over SCTP.
The user can input messages to send to the server, and "exit" can be used to close the client's connection.
It sends messages to the server over SCTP for communication.
Key Points:

Socket Creation: The client creates a socket for communication with the server.
Connection Establishment: It establishes a connection with the server using the specified server IP address and port.
User Interaction: The client interacts with the user, allowing them to input messages to send to the server.
Message Sending: Messages entered by the user are sent to the server using SCTP.
Usage of the Code:

The code can be compiled and executed as an SCTP client.
Upon execution, the client connects to the specified server.
The user can input messages, which will be sent to the server.
Typing "exit" allows the client to close its connection with the server.
This code represents a simple SCTP client that can establish a connection to a remote server and send messages to it. It provides a basic example of how to interact with an SCTP server for communication.